Remove obsolete code, fix test
